Re -> mu, vtk issues
Re changes with fluid velocity, so Brinkman flow is now specified via dynamic viscosity instead of Re.
There are apparently serious issues with VTK on MacOSX which result in a system error that immediately crashes Python upon loading the VTK library. After hours of looking, I can find no fix for any version of Python 3.x. To be able to run unit tests on this platform, the ModuleNotFoundError will be caught and handled, and vtk tests are skipped when vtk is not installed.