starts adding swarm dataframe, breaks code
This commit undoubtedly breaks everything, but it is a starting place.
swarm now has a dataframe that contains random walk mu and cov by default.
The default update_positions method now expects these.
gaussian_walk has been updated to recognize when the random walk parameters are all the same or different. If they are the same, do a quick vectorized call to RndState. Otherwise, enter a for loop and get an appropriate value for each agent.
No testing or updating of examples and other code has been done.