Some bug fixes, but marking FTLE as broken
Results do not match up with VisIt and I know not why. I'm questioning
even the RK45 solver at this point. A careful debug will require better
visualization of the flow field to compare with particle movement, and
maybe plotting of integral curves. So these are going to go up in
priority and FTLE will have to languish for a while.