New RKDP solver replaces RKF, bug fix in FTLE
This major commit scraps the RK45 solver completely (I think it was
still somewhat buggy) in favor of a Dormand-Prince solver.
Unfortunately, since boundary conditions are not checked inside the
solver, the FSAL property will only help for the advection steps, but
the overall simplicity of the code in the rewrite is a huge step
forward. There is also a fix in the FTLE solver where division is now by
the full integration time always, instead of the first time a pixel in
the stencil leaves the domain. This solved certain edge-of-domain
problems.