Closes #30: Implement FTLE
This is yet to be tested for 3D cases, it only calculuates one FTLE
field at a time for now (mesh of time points is TODO), and it's unusably
slow when there's an immersed mesh structure to take into account. But
it runs, appears to give good, consistent results in 2D, and is fully
generalized!