Sets up visual test mvib2d. Begins addressing zero issues
An immediate issue for the 2d moving boundary interaction routine is the
potential for divide by zero issues in closest_dist_btwn_two_lines. This
commit begins to work on this by setting up preliminary checks in that
function to catch cases where two vertex points are very close together,
and geometric calculations in geom to deal with it. We also need to
catch the case where an agent moves very little if at all, so a0 and a1
are essentially the same point.