ibmesh roundoff error bug fixed
The ibmesh was being stored as an array of float32, probably because the
source was float32. The resulting roundoff error in interpolation
between meshpoints was allowing penetration by agents when collision
happens very close to the beginning or end of a timestep - such as in
the case of an agent that had previously collided and was facing
immediate collision again. The ibmesh is now converted to float64 right
before being stored as an attribute. This was sufficient to solve the
original problem.