pos_history now updated after positions
Previously, a copy of positions was being put in pos_history, and then
get_positions was called to update positions. This could cause serious
problems if get_positions is non Markovian and needs to know the prior
positions of the swarm! Now the current position of the swarm is copied,
and pos_history is updated after get_positions has been called (but
before applying boundary conditions).