plot_all has headings (but warning), vel_history added, bug fix
The vel_history (and full_vel_history) attribute was added to swarm and implemented in move.
This means that we are now recording the full velocity history in parallel to the position
history. This was necessary in order to easily plot heading indicators for all time after
the simulation has run, but it is probably also just good practice since it doesn't take
much memory and could easily be used for data analysis.

plot_all heading indicators now seem to work, at least when they are based on the vel_history.
Testing needs to be done to see if it will also work with an 'angles' property. The code is
supposed to warn if the angles property was not stored in props_history and then fall back
on using the vel_history. There is also a warning about a nan conversion having to do with
a masked array that needs exploring and squashing.

This commit also includes a bug fix where cir_rad was hard coded in plot. It would be worth
testing to make sure that the new cir_rad parameter in plot() and plot_all() works as intended.