Major update to projection; proj_end bug fix
Previously, an agent intersects a mesh element, and then the point of
intersection would update to remain on the same location on the mesh
element as the element moved. Added to this was projected agent
movement. But updating the base point like that implies an infinite
frictional force as the element moves laterally even as we are assuming
a frictionless surface for agent movement. As a result, we need to
update the point of intersection but subtract off lateral movement so
that the point of intersection is only updated ortho to the element.

proj_end was previously not operating on a normalized Q_end, resulting
in an incorrect direction.