Fixes incorrect math in rotation detection
Separation between the agent and the mesh element occurs whenever the
speed of the mesh element is greater than the speed of the agent in the
direction orthogonal to the mesh element. This is now tested for, and if
it occurs within the time step and before an edge is reached, we find
the time of separation using the brentq root finding algorithm. The
notes have been updated accordingly.