Fixes acute geometry bug, EPS in wrong dir, rubberband PASSING!
Splitting geometry into acute versus obtuse cases doesn't work because
the agent could be sliding in one direction while the movement of the
mesh past it pushes it in the opposite direction in the Lagrangian
frame. Instead, we let recursion do the work and catch repeat cases on
the same mesh element from an adjoining one. Also, the EPS when an agent
ends in a corner needed to be in a direction that bisects the angle
between the mesh elements, otherwise penetration could occur. All of
this has been fixed, and the first rubberband simulation shows no
penetration.